feat(borrow): extract quote logic and success component
