+++ src/components/BorrowTab.tsx
@@ -1,26 +1,23 @@
 "use client";
 
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import { useQuery } from "@tanstack/react-query";
-import Image from "next/image";
 import { useRouter } from "next/navigation";
+import Image from "next/image";
 import styles from "./BorrowTab.module.css";
-import Button from "./Button"; // Reuse Button component
-import { Asset } from "@/types/common";
-import { FormattedRuneAmount } from "./FormattedRuneAmount"; // Reuse FormattedRuneAmount
+import Button from "./Button";
 import CollateralInput from "./CollateralInput";
 import BorrowQuotesList from "./BorrowQuotesList";
+import BorrowSuccessMessage from "./BorrowSuccessMessage";
+import { Asset } from "@/types/common";
+import { FormattedRuneAmount } from "./FormattedRuneAmount";
 import { useBorrowProcess } from "@/hooks/useBorrowProcess";
+import useBorrowQuotes from "@/hooks/useBorrowQuotes";
 import {
-  fetchPopularFromApi,
   fetchRuneBalancesFromApi,
   fetchRuneInfoFromApi,
   fetchRuneMarketFromApi,
-  fetchBorrowQuotesFromApi,
-  fetchBorrowRangesFromApi,
   QUERY_KEYS,
-  LiquidiumBorrowQuoteResponse,
-  LiquidiumBorrowQuoteOffer,
 } from "@/lib/apiClient";
 import { normalizeRuneName } from "@/utils/runeUtils";
 import {
@@ -60,93 +57,9 @@ export function BorrowTab({
   signPsbt,
 }: BorrowTabProps) {
   const router = useRouter();
-  // --- State ---
   const [collateralAsset, setCollateralAsset] = useState<Asset | null>(null);
-  const [collateralAmount, setCollateralAmount] = useState(""); // Amount of Rune to use as collateral
-  // State for rune fetching/searching
-  const [isPopularLoading, setIsPopularLoading] = useState(false); // Local state for popular runes
-  const [popularRunes, setPopularRunes] = useState<Asset[]>([]);
-  const [popularError, setPopularError] = useState<string | null>(null);
+  const [collateralAmount, setCollateralAmount] = useState("");
 
-  // State for quotes
-  const [quotes, setQuotes] = useState<LiquidiumBorrowQuoteOffer[]>([]);
-  const [isQuotesLoading, setIsQuotesLoading] = useState(false);
-  const [quotesError, setQuotesError] = useState<string | null>(null);
-  const [selectedQuoteId, setSelectedQuoteId] = useState<string | null>(null);
-  const [minMaxRange, setMinMaxRange] = useState<string | null>(null);
-  const [borrowRangeError, setBorrowRangeError] = useState<string | null>(null);
-
-  // --- Data Fetching ---
-  // Fetch popular runes on mount using API
-  useEffect(() => {
-    const fetchPopular = async () => {
-      setIsPopularLoading(true);
-      setPopularError(null);
-      setPopularRunes([]);
-      try {
-        // Define the hardcoded asset
-        const liquidiumToken: Asset = {
-          id: "liquidiumtoken",
-          name: "LIQUIDIUM•TOKEN",
-          imageURI: "https://icon.unisat.io/icon/runes/LIQUIDIUM%E2%80%A2TOKEN",
-          isBTC: false,
-        };
-        const response = await fetchPopularFromApi();
-        let mappedRunes: Asset[] = [];
-        if (!Array.isArray(response)) {
-          mappedRunes = [liquidiumToken];
-        } else {
-          const fetchedRunes: Asset[] = response
-            .map((collection: Record<string, unknown>) => ({
-              id: (collection?.rune_id as string) || `unknown_${Math.random()}`, // Use rune_id if available
-              name: (
-                (collection?.slug as string) ||
-                (collection?.rune as string) ||
-                "Unknown"
-              ).replace(/-/g, "•"), // Use slug or rune name
-              imageURI:
-                (collection?.icon_content_url_data as string) ||
-                (collection?.imageURI as string),
-              isBTC: false,
-            }))
-            .filter(
-              (rune) =>
-                rune.id !== liquidiumToken.id &&
-                normalizeRuneName(rune.name) !==
-                  normalizeRuneName(liquidiumToken.name),
-            );
-          mappedRunes = [liquidiumToken, ...fetchedRunes];
-        }
-        setPopularRunes(mappedRunes);
-        // Set default collateral asset if none is selected
-        if (!collateralAsset && mappedRunes.length > 0) {
-          setCollateralAsset(mappedRunes[0]);
-        }
-      } catch (error) {
-        setPopularError(
-          error instanceof Error
-            ? error.message
-            : "Failed to fetch popular runes",
-        );
-        const liquidiumTokenOnError: Asset = {
-          id: "liquidiumtoken",
-          name: "LIQUIDIUM•TOKEN",
-          imageURI: "https://icon.unisat.io/icon/runes/LIQUIDIUM%E2%80%A2TOKEN",
-          isBTC: false,
-        };
-        setPopularRunes([liquidiumTokenOnError]);
-        if (!collateralAsset) {
-          setCollateralAsset(liquidiumTokenOnError);
-        }
-      } finally {
-        setIsPopularLoading(false);
-      }
-    };
-    fetchPopular();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []); // Run only once
-
-  // Query for Rune Balances
   const { data: runeBalances, isLoading: isRuneBalancesLoading } = useQuery<
     OrdiscanRuneBalance[],
     Error
@@ -157,7 +70,6 @@ export function BorrowTab({
     staleTime: 30000,
   });
 
-  // Query for Collateral Rune Info (for decimals)
   const { data: collateralRuneInfo, isLoading: isCollateralRuneInfoLoading } =
     useQuery<RuneData | null, Error>({
       queryKey: [QUERY_KEYS.RUNE_INFO, collateralAsset?.name],
@@ -169,7 +81,6 @@ export function BorrowTab({
       staleTime: Infinity,
     });
 
-  // Query for Collateral Rune Market Info (for USD value)
   const { data: collateralRuneMarketInfo } = useQuery<
     OrdiscanRuneMarketInfo | null,
     Error
@@ -180,7 +91,27 @@ export function BorrowTab({
         ? fetchRuneMarketFromApi(collateralAsset.name)
         : Promise.resolve(null),
     enabled: !!collateralAsset && !collateralAsset.isBTC,
-    staleTime: 5 * 60 * 1000, // 5 minutes
+    staleTime: 5 * 60 * 1000,
+  });
+
+  const {
+    popularRunes,
+    isPopularLoading,
+    popularError,
+    quotes,
+    isQuotesLoading,
+    quotesError,
+    selectedQuoteId,
+    setSelectedQuoteId,
+    minMaxRange,
+    borrowRangeError,
+    resetQuotes,
+    handleGetQuotes,
+  } = useBorrowQuotes({
+    collateralAsset,
+    collateralAmount,
+    address,
+    collateralRuneInfo: collateralRuneInfo ?? null,
   });
 
   const {
@@ -200,84 +131,6 @@ export function BorrowTab({
     collateralRuneInfo: collateralRuneInfo ?? null,
   });
 
-  // Fetch min-max range when collateral asset changes
-  useEffect(() => {
-    const fetchMinMaxRange = async () => {
-      if (!collateralAsset || !address || collateralAsset.isBTC) {
-        setMinMaxRange(null);
-        setBorrowRangeError(null);
-        return;
-      }
-
-      try {
-        // Get the actual rune ID from collateralRuneInfo if available
-        let runeIdForApi = collateralAsset.id;
-
-        if (collateralRuneInfo?.id?.includes(":")) {
-          runeIdForApi = collateralRuneInfo.id;
-        }
-
-        const result = await fetchBorrowRangesFromApi(runeIdForApi, address);
-
-        if (result.success && result.data) {
-          const { minAmount, maxAmount } = result.data;
-
-          // Convert raw values to formatted values based on decimals
-          const decimals = collateralRuneInfo?.decimals ?? 0;
-          const minFormatted = formatRuneAmount(minAmount, decimals);
-          const maxFormatted = formatRuneAmount(maxAmount, decimals);
-
-          setMinMaxRange(`Min: ${minFormatted} - Max: ${maxFormatted}`);
-          setBorrowRangeError(null);
-        } else {
-          setMinMaxRange(null);
-          setBorrowRangeError(null);
-        }
-      } catch (error) {
-        console.error("[BorrowTab] Error fetching min-max range:", error);
-        const errorMessage =
-          error instanceof Error ? error.message : String(error);
-        setMinMaxRange(null);
-        // Detect "No valid ranges found" error
-        if (
-          errorMessage.includes("No valid ranges found") ||
-          errorMessage.includes(
-            "Could not find valid borrow ranges for this rune",
-          )
-        ) {
-          setBorrowRangeError(
-            "This rune is not currently available for borrowing on Liquidium.",
-          );
-        } else {
-          setBorrowRangeError(null);
-        }
-      }
-    };
-
-    fetchMinMaxRange();
-  }, [collateralAsset, address, collateralRuneInfo]);
-
-  // --- Helper Functions ---
-  // Helper function to format rune amounts with BigInt precision
-  const formatRuneAmount = (rawAmount: string, decimals: number): string => {
-    try {
-      // Using BigInt to maintain precision
-      const rawAmountBigInt = BigInt(rawAmount);
-      const divisorBigInt = BigInt(10 ** decimals);
-
-      // Scale by 100 for two decimal places of precision
-      const scaledAmount = (rawAmountBigInt * BigInt(100)) / divisorBigInt;
-
-      // Convert to number for formatting (safe now that we've scaled down)
-      const scaledNumber = Number(scaledAmount) / 100;
-
-      return scaledNumber.toFixed(decimals > 0 ? 2 : 0);
-    } catch {
-      // Fallback to basic number conversion if BigInt fails
-      return (Number(rawAmount) / 10 ** decimals).toFixed(decimals > 0 ? 2 : 0);
-    }
-  };
-
   const getSpecificRuneBalance = (
     runeName: string | undefined,
   ): string | null => {
@@ -287,125 +140,14 @@ export function BorrowTab({
     return found ? found.balance : "0";
   };
 
-  // --- Handlers ---
   const handleSelectCollateral = (asset: Asset) => {
     setCollateralAsset(asset);
     setCollateralAmount("");
-    setQuotes([]);
-    setQuotesError(null);
+    resetQuotes();
     setSelectedQuoteId(null);
     resetLoanProcess();
-    setMinMaxRange(null);
-    setBorrowRangeError(null); // Clear any existing borrow range error
   };
 
-  const handleGetQuotes = async () => {
-    if (!collateralAsset || !collateralAmount || !address) {
-      return;
-    }
-
-    setIsQuotesLoading(true);
-    setQuotesError(null);
-    setQuotes([]);
-    setSelectedQuoteId(null);
-    resetLoanProcess();
-
-    try {
-      // Convert user amount to raw amount based on decimals
-      const decimals = collateralRuneInfo?.decimals ?? 0;
-
-      // Calculate raw amount with proper decimal handling
-      let rawAmount;
-      try {
-        // Pure BigInt calculation to maintain precision
-        const amountFloat = parseFloat(collateralAmount);
-        // Convert to integer representation (e.g. 1.23 -> 123 for 2 decimals)
-        const amountInteger = Math.floor(
-          amountFloat * 10 ** Math.min(8, decimals),
-        );
-        // Scale to full decimal precision
-        const multiplier = BigInt(10 ** Math.max(0, decimals - 8));
-        const amountBigInt = BigInt(amountInteger) * multiplier;
-        rawAmount = amountBigInt.toString();
-      } catch {
-        // Fallback calculation
-        rawAmount = String(
-          Math.floor(parseFloat(collateralAmount) * 10 ** decimals),
-        );
-      }
-
-      // Get the actual rune ID from collateralRuneInfo if available
-      // This ensures we're using the correct ID format (e.g., "810010:907")
-      let runeIdForApi = collateralAsset.id;
-
-      if (collateralRuneInfo?.id?.includes(":")) {
-        runeIdForApi = collateralRuneInfo.id;
-      }
-
-      const result: LiquidiumBorrowQuoteResponse =
-        await fetchBorrowQuotesFromApi(
-          runeIdForApi, // Use the correct rune ID format for the API call
-          rawAmount,
-          address,
-        );
-
-      if (result?.runeDetails) {
-        // Extract min-max range if available
-        if (result.runeDetails.valid_ranges?.rune_amount?.ranges?.length > 0) {
-          const ranges = result.runeDetails.valid_ranges.rune_amount.ranges;
-
-          // Find global min and max across all ranges
-          let globalMin = BigInt(ranges[0].min);
-          let globalMax = BigInt(ranges[0].max);
-
-          // Compare with other ranges to find global min and max
-          for (let i = 1; i < ranges.length; i++) {
-            const currentMin = BigInt(ranges[i].min);
-            const currentMax = BigInt(ranges[i].max);
-
-            if (currentMin < globalMin) globalMin = currentMin;
-            if (currentMax > globalMax) globalMax = currentMax;
-          }
-
-          // Convert raw values to formatted values based on decimals
-          const decimals = collateralRuneInfo?.decimals ?? 0;
-          const minFormatted = formatRuneAmount(globalMin.toString(), decimals);
-          const maxFormatted = formatRuneAmount(globalMax.toString(), decimals);
-
-          setMinMaxRange(`Min: ${minFormatted} - Max: ${maxFormatted}`);
-        } else {
-          setMinMaxRange(null);
-        }
-
-        // Process offers
-        if (result.runeDetails.offers) {
-          setQuotes(result.runeDetails.offers);
-          if (result.runeDetails.offers.length === 0) {
-            setQuotesError("No loan offers available for this amount.");
-          }
-        } else {
-          setQuotes([]);
-          setQuotesError("No loan offers found or invalid response.");
-        }
-      } else {
-        setQuotes([]);
-        setQuotesError("No loan offers found or invalid response.");
-        setMinMaxRange(null);
-      }
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : "Failed to fetch quotes.";
-      setQuotesError(errorMessage);
-      setQuotes([]);
-      setMinMaxRange(null);
-    } finally {
-      setIsQuotesLoading(false);
-    }
-  };
-
-  // Asset selector is now handled by the InputArea component
-
-  // --- Render Logic ---
   const isLoading = isQuotesLoading || isPreparing || isSigning || isSubmitting;
   const canGetQuotes =
     connected &&
@@ -444,7 +186,6 @@ export function BorrowTab({
     <div className={styles.borrowTabContainer}>
       <h1 className="heading">Borrow Against Runes</h1>
 
-      {/* Collateral Input Area */}
       <CollateralInput
         connected={connected}
         collateralAsset={collateralAsset}
@@ -452,9 +193,8 @@ export function BorrowTab({
         collateralAmount={collateralAmount}
         onCollateralAmountChange={(value) => {
           setCollateralAmount(value);
-          setQuotes([]);
+          resetQuotes();
           setSelectedQuoteId(null);
-          setQuotesError(null);
         }}
         availableAssets={popularRunes}
         isAssetsLoading={isPopularLoading}
@@ -465,42 +205,32 @@ export function BorrowTab({
         minMaxRange={minMaxRange || undefined}
         onPercentageClick={(percentage) => {
           if (!connected || !collateralAsset) return;
-
           const rawBalance = getSpecificRuneBalance(collateralAsset.name);
           if (!rawBalance) return;
-
           const balanceNum = parseFloat(rawBalance);
           if (isNaN(balanceNum)) return;
-
           const decimals = collateralRuneInfo?.decimals ?? 0;
           const availableBalance = balanceNum / 10 ** decimals;
-
           const newAmount =
             percentage === 1 ? availableBalance : availableBalance * percentage;
-
           const formattedAmount =
             Math.floor(newAmount * 10 ** decimals) / 10 ** decimals;
-
           setCollateralAmount(formattedAmount.toString());
-          setQuotes([]);
+          resetQuotes();
           setSelectedQuoteId(null);
-          setQuotesError(null);
         }}
       />
 
-      {/* Borrow Range Error */}
       {borrowRangeError && (
         <div className="errorText" style={{ marginBottom: 8 }}>
           {borrowRangeError}
         </div>
       )}
 
-      {/* Get Quotes Button */}
       <Button onClick={handleGetQuotes} disabled={!canGetQuotes || isLoading}>
         {isQuotesLoading ? "Fetching Quotes..." : "Get Loan Quotes"}
       </Button>
 
-      {/* Display Quotes */}
       {quotesError && <div className="errorText">{quotesError}</div>}
       {quotes.length > 0 && (
         <BorrowQuotesList
@@ -510,7 +240,6 @@ export function BorrowTab({
         />
       )}
 
-      {/* Start Loan Button */}
       {selectedQuoteId && (
         <Button
           onClick={() => startLoan(selectedQuoteId, collateralAmount)}
@@ -526,7 +255,6 @@ export function BorrowTab({
         </Button>
       )}
 
-      {/* Display Loan Process Status/Error/Success */}
       {loanProcessError && (
         <div className={`errorText ${styles.messageWithIcon}`}>
           <Image
@@ -539,45 +267,18 @@ export function BorrowTab({
           <span>Error: {loanProcessError}</span>
         </div>
       )}
-      {loanTxId && (
-        <div className={`${styles.messageWithIcon} ${styles.successMessage}`}>
-          <Image
-            src="/icons/check-0.png"
-            alt="Success"
-            className={styles.messageIcon}
-            width={16}
-            height={16}
-          />
-          <div className={styles.successContent}>
-            <p className={styles.successTitle}>Loan started successfully!</p>
-            <p className={styles.successDetails}>
-              Your loan has been initiated. It will appear in your portfolio
-              shortly.
-            </p>
-            <div className={styles.successButtons}>
-              <Button
-                onClick={() => {
-                  // Navigate to the portfolio tab using client-side routing
-                  router.push("/?tab=portfolio", { scroll: false });
-                }}
-                style={{ marginRight: "var(--space-2)" }}
-              >
-                View Portfolio
-              </Button>
-              <Button
-                onClick={() => {
-                  // Reset the loan process to start a new loan
-                  resetLoanProcess();
-                  setSelectedQuoteId(null);
-                  setQuotes([]);
-                }}
-              >
-                Start Another Loan
-              </Button>
-            </div>
-          </div>
-        </div>
-      )}
+
+      <BorrowSuccessMessage
+        loanTxId={loanTxId}
+        onViewPortfolio={() =>
+          router.push("/?tab=portfolio", { scroll: false })
+        }
+        onStartAnother={() => {
+          resetLoanProcess();
+          setSelectedQuoteId(null);
+          resetQuotes();
+        }}
+      />
     </div>
   );
 }
